pkgDepCache::MarkInstallRec(): just declare Depth as int, and get rid of a cast
Anyway, it's used by printf, and it expects an int.

This ABI change must not be noticeable because this method is
* protected (hence not used elsewhere);
* non-virtual (hence no effect on VMT).

+++ apt-pkg/depcache.h
@@ -217,7 +217,7 @@ class pkgDepCache : protected pkgCache::Namespace
    // implementation
    void MarkInstallRec(PkgIterator const &Pkg,
       bool Restricted, std::set<PkgIterator> &MarkAgain,
-      unsigned long Depth, const char *DebugStr);
+      int Depth, const char *DebugStr);
 
    public:
    void SetReInstall(PkgIterator const &Pkg,bool To);
